#298 refactor integration tests
