+++ extension/test/integration/get-payment-methods.handler.spec.js
@@ -13,7 +13,7 @@ describe('::getPaymentMethods::', () => {
   })
 
   after(async () => {
-    await iTSetUp.cleanupResources(ctpClient)
+    await iTSetUp.stopRunningServers()
   })
 
   it('given a payment ' +
