+++ extension/test/integration/make-payment.handler.spec.js
@@ -13,7 +13,7 @@ describe('::makePayment::', () => {
   })
 
   afterEach(async () => {
-    await iTSetUp.cleanupResources(ctpClient)
+    await iTSetUp.stopRunningServers()
   })
 
   it('given a payment ' +
@@ -70,46 +70,5 @@ describe('::makePayment::', () => {
     expect(transaction.amount.currencyCode).to.be.equal('EUR')
     expect(transaction.amount.centAmount).to.be.equal(paymentDraft.amountPlanned.centAmount)
     expect(transaction.interactionId).to.be.a('string')
-  })
-
-  it('given a payment,' +
-    'when makePayment custom field is set with Klarna Adyen request without line items,' +
-    'then it should calculate correct line items for Klarna Adyen', async () => {
-    const payment = await iTSetUp.initPaymentWithCart(ctpClient)
-    const makePaymentRequestDraft = {
-      riskData: {
-        clientData: 'eyJ2ZXJzaW9uIjoiMS4wLjAiLCJkZXZpY2VGaW5nZXJwcmludCI6ImRmLXRpbWVkT3V0In0='
-      },
-      reference: 'Klarna Pay later',
-      paymentMethod: {
-        type: 'klarna'
-      },
-      amount: {
-        currency: 'EUR',
-        value: '8610'
-      },
-      shopperLocale: 'de_DE',
-      countryCode: 'DE',
-      shopperEmail: 'youremail@email.com',
-      shopperReference: 'YOUR_UNIQUE_SHOPPER_ID',
-      returnUrl: 'https://www.yourshop.com/checkout/result'
-    }
-
-    const { statusCode, body: updatedPayment } = await ctpClient.update(ctpClient.builder.payments,
-      payment.id, payment.version, [
-        {
-          action: 'setCustomField',
-          name: 'makePaymentRequest',
-          value: JSON.stringify(makePaymentRequestDraft)
-        }
-      ])
-
-    expect(statusCode).to.be.equal(200)
-    const makePaymentInteraction = updatedPayment.interfaceInteractions[0].fields
-    const makePaymentRequest = JSON.parse(makePaymentInteraction.request)
-    const makePaymentResponse = JSON.parse(makePaymentInteraction.response)
-
-    expect(makePaymentRequest.lineItems).to.have.lengthOf(3)
-    expect(makePaymentResponse.resultCode).to.be.equal('RedirectShopper')
   })
 })
